Fix deadlock in controller/migration interaction
This works around a nasty, complicated issue. Previously, domains would
send two messages as part of their bootup process: one to the `Blender`
on the controller (via their internal control channel), and one to the
distributed Soup coordination logic via the worker <-> controller
connection. This produced a subtle race and deadlock because the
migration code (running `Blender` logic) runs in a different thread to
the coordination logic. As a result, the `Blender`'s migration code
might wait for a domain to be ready to start processing writes while the
domain was blocked waiting for its connectivity information to show up
in the worker's local `ChannelCoordinator`. This informatio never showed
up because the controller's coordination thread was trying to acquire
the lock on the `Blender` in order to insert information about read
endpoints that it received as part of the `DomainBooted` control-plane
message. The problem only shows up with >=2 physical machines on a real
network (presumably because loopback doesn't produce this race).

The current solution moves the connectivity gossip in response to a
domain boot into the `Blender`; this further muddles the
`Controller`/`Blender` divide and requires a bunch of annoying
`clone()`'s, unfortunately, but lets us make progress.