Mixed-key batches on mutator connections
This sends batches of writes to base node shards instead of sending each
write individually (which we previously did in order to match the
semantics of single-machine mutators' channels). This yields a massive
improvement in throughput: a single writer client can now do ~2M PUTs
per second (previously: ~500k); with 10 shards over 10 machines, we max
out at 3.5M PUTs/sec without further tuning.

The implementation is a bit haphazard and does some allocation and
cloning that may not be necessary, although the costs of syscalls dwarfs
this anyway. (The main reason is that `DomainInputHandle::base_send`
already receives a `Packet` so that it can be generic over TX and non-TX
writes; the sharding code now has to break apart, clone and reassemble
`Packets`.)