Fix lots of bugs at once
The previous way of checking for partial and adding indexes was very
broken. This was in part because it only looked at one domain at a time,
but also because it didn't iterate over the nodes in any predictable
order. This in turn meant that one view may add a replay obligation for
another view, but that view never realized, nor did it propagate that
obligation to its ancestors, etc.