Fix reuse query corruption bug
When we extend an existing query, we first traverse the old query to
find matching operators in the new one; then we traverse the new query
and rewrite it, replacing the new MIR nodes with Reuse nodes for the
reused operators from the old query.

This process was buggy: when rewiring the children of a MIR node, we
would afterwards iterate not over the original children in the new
query, but rather over the newly written ones, which could include ones
that didn't exist in the original query (e.g., from other sibling
queries). This resulted in corrupt queries being added in some rare
cases (!), and apparently survived without detection for five months
(!!). Whoopsy.