Make migrations happen in a closure
I've been wanting to make this change for a while now.
Since a migration mutably borrows the `Blender`, the migration usually
has to happen in its own (useless) scope, at least without NLL. This is
annoying. Furthermore, the requirement to call `mig.commit()` is silly
since we (at least atm) cannot abort a migration.

This patch changes migrations to instead happen in a closure, which
essentially wraps the same things every caller had to do previously.