Disentangle GraphViz and debug output for MIR
This is a long-standing cleanup: we previously used GraphViz format
output for the implementations of `Debug` and `Display` on various MIR
data structures, neither did so consistently nor included all the
necessary information.

With this change, there is a new trait, `mir::visualize::GraphViz` that
allows getting all information in valid GraphViz format, and the other
implementations are restored to something more akin to human-readable
debug output.