Move controller thread creation outside build()
This is necessary because we otherwise end up back in the situation
where a `ControllerInner` is sent across thread boundaries (namely,
`self` is sent into the `thread::spawn` closure).
3aebfe8b19557bf4be14c5358e76d2938388fbf1 laid the ground work for
this in order to support long-term recipe storage in `ControllerInner`,
but this change is required before we can actually store a recipe.