Store current recipe across migrations again
This brings back storage of the current recipe, and accordingly, of the
`SqlIncorporator` it contains. Since the `SqlIncorporator` contains all
cached state related to query reuse, this effectively re-enables reuse.
(Previously, reuse could be enabled with the refactored code, but would
not actually happen.)

This change also fixes the workaround developed for the "security-merge"
branch in 2cbd8c3ad4ca44762ef8dbfa8455a9bfec41d242, which is now no
longer required (@laratimbo).