Remove redundant leaf materializations
Previously, we would add a leaf materialization for every query added to
the graph, including subqueries of compound queries and nested queries.
This makes no sense as these aren't named queries that the user would
have any way of actually reading the results of -- so it's a waste to
expend space and time on maintaining them. This change removes readers
(and thus, extra domains and materializations) for all "internal"
queries, but maintains them for any explicitly specified user query in
the recipe.

Note that this is still correct if the user changes their mind later, or
incidentally writes the same query: we will pick this up via our reuse
logic and simply add a reader to the existing query.

For reasons that aren't entirely clear to me, this also had the effect
of fixing the post-migration crash in the "vote" benchmark (?).