Broken no-conflict policy in SecureCRP
First attempt at writing a policy that deals with paper conflicts. Does
not currently work because we do not support NOT IN, and adding support
isn't trivial: we implement IN as a JOIN against the subquery results,
and NOT IN (= "except"/"unless") can't trivially be modeled as a JOIN.
A LEFT JOIN followed by a filter to keep only non-matching rows (i.e.,
those with a NULL value in a right relation column) and a projection
might work, however.