Remove VoteCount as a public edge view in "vote"
This makes VoteCount a nested query in the data-flow, and consequently
removes the leaf materialization for it. Doing so saves space and
resources (one fewer domain, and less work per update).

This still works even with migrations, since the new queries are
(perhaps idiosyncratically) allowed to refer to named nested queries in
existing queries. To fix this in a more principled way, we probably want
to introduce a notion of public leaf views and internal shorthand views.