Don't materialize all compound queries' leaves
Similar to earlier changes, this removes extra, redundant leaf
materializations on compound queries if they aren't top-level queries
that the recipe requests to be materialized.

N.B.: doesn't quite handle the case of a UNION with a TOP-K correctly,
as the Top-K node would need to sanitize leaf columns if it is indeed
the final node in the query.