Fix vote-stress benchmark
Ported "vote-stress" to the refactored APIs. Now compiles and runs
again, although it requires the migration to run in "--stupid" mode as
the smart, reusing migration fails with a full-above-partial error (the
union is being materialized?!). Hence, performance is likely to be poor.

Note that "vote-stress" also does single-element `mput` invocations,
which is slow (albeit it's not being measured here).