Fix some bugs
Fix an annoying reuse bug that occured when the same node is
reused by more than one node in the new query. A better fix
for this would be to have `optimize` deal with in-query reuse,
but this is simpler and forces each node to be reused only once.

Fix another bug in graph construction where a table with no
policies would leave dangling nodes.