Add additive "extend_recipe" RPC
The only prior route to change the recipe was the "install_recipe" RPC,
which burdens the caller with maintaining the recipe state and requires
the whole recipe to be repeated if a query is to be added. The new RPC
allows strict *addition* to a recipe, and merely re-exposes an API
(`Recipe::extend`) that always existed.