Restore failing partial test w/ empty parents
Should fix this. Specifically, we should not eagerly fully materialize
nodes below empty parents since while creating the full materialization
is free, its maintenance isn't.

Instead, we should simply not replay if one input to a join on the
replay path for a new materialization is empty (except if it's a left
join and the right side is empty, of course).