Disable empty parent optimization
This optimization forces full materialization of known-empty views even
if we earlier decided to make them partial (and continue to claim that
they are). Empty bases can still skip the replay; we'll probably want
some more sophisticated logic that also avoids replays if we know that
a node necessarily starts off empty (e.g., because a parent joins an
empty view).

Fixes RTM's strange partial-does-not-discard bug. We pass the new test
now also.