Fix behavior of partial across shuffles
Previously, a partial replay where the source was sharded would only
result in a single partial replay request. However, that is not always
the correct approach. If the source is sharded by a *different* key
column than the one being requested, we need to ask *all* the shards!

Furthermore, *only* in the situation where we ask all the shards should
the shard merger buffer replay responses. When only one shard was asked,
there is no need for buffering!

This commit fixes that behavior.