Ensure that replays always reach querying shard
Imagine a graph with two nodes, A and B. B is sharded, so there's a
sharder between them. B is partial over some column from A that is *not*
B's shard key. B.1 requests key k1 from A. B.2 requests key k2 from A. A
batches them together (remember batched replays), and sends a single
replay response with k1+k2. The sharder now needs to route that packet.
What does it do? It doesn't know that A.1 requested k1 and A.2 requested
k2. So, it needs to forward the entire replay to both shards (and *all*
shards in the general case). This extends to the case where a particular
sharded domain has *no* materializations: now, the replays the sharder
sees could potentially come from someone *downstream* of the B shards,
so it definitely must send to all shards.

We *may* be able to make things a little better by detecting when B's
shard key is the same as the lookup key for the partial. In that case,
I think we know that only one shard *could have* asked for it. But we
should treat that as an optimization.

There is still an issue with tests occasionally failing due to
duplicated records. I'm not sure why yet.