Don't crash for mutators for non-existent bases
The RPCs to get a write handle could not previously return nothing or
indicate an error, so Soup would crash if a write handle for a
non-existent table was requested. We now return an `Option` and expect
the client application to handle the error.

We might wish to unify this with the use of `RpcError` in other places
to indicate to the application what the problem was.