Expose column schema on Getters
Read handles for views need to carry a representation of their column
schema, so that the MySQL shim can correctly piece together the response
to the client. Ultimately, we will need to expose SQL types and flags
also, but this is a quick hack that only relies on information currently
available in the data-flow graph.