Include bogokeys in expected test output
For queries without parameters, we now include a bogokey (see 14e9adf).
Many of our test queries are of this form, but the tests' expected edge
view schema did not include the bogokey. Filtering it and all of its
effects out systematically turned out to be tricky, so I just included
it for now. In the long run, we'll want a `read_all` RPC with similar
effect to a bogokey lookup, because that will allow us to get rid of
this silliness altogether.