Expose state size via get_statistics HTTP endpoint
We already have support for obtaining statistics from a domain via the
`DomainStatistics` structure returned via a `GetStatistics` payload
packet. This change adds the memory used for state associated with each
node in the domain to the `NodeStats` part of the domain statistics.

Note that the size calculation for `State` is a lower bound: it only
includes the row data once (even if there are multiple indices; we'll
fix this shortly), and it doesn't include the size of the hash table
structure itself.