Clarify semantics of QG signatures and fix test
We've for a long time had some bugs in the reuse algorithm:
specifically, ExactMatch would trigger even for queries that differ in
literals or operators in their predicates. The reason for this was a
sloppy use of the semantics of `QueryGraphSignature`: we treated it
(and, specifically, `QueryGraphSignature::hash`) as a means of telling
whether two queries are *identical* (i.e., matching hashes => matching
queries), but the signature really only has much weaker semantics.
Matching signatures indicate that the QGs share the same relation and
that their predicates mention the same attributes (= columns), as per
Finkelstein paper. However, the caller is still obliged to check that
the predicates in the queries are actually identical.

In debugging this, @larat7 and myself added some tests that assumed
stronger semantics; this commit fixes them to rely on the correct
semantics of QG signatures. Following commits will fix the actual bug.