Small, but important fix for identical queries
`ExactMatch` reuse of identical queries directly aliases a new query
onto the existing query's leaf node, and thus avoids going through the
MIR layer. Because we now build getters based on the query name,
however, it's crucial that we also short-circuit the logic that finds
the leaf `NodeIndex` for a named query. Previously, this logic would end
up recursing into the `MirToSqlConverter`, which had never heard of the
aliased query!