Max memory arg to Souplet; thread through layers
This adds a new argument to the "souplet" binary: --memory or -m. This
will (in the future) have the effect of setting a maximum aggregate size
in bytes for all materializations on this machine.

So far, we only thred the information through the layers, so that it is
available from within the Souplet. Next step: use GetStatistics RPCs on
local domains to get state size, compute estimate over time.