Regularly check aggregate state size at worker
Adds a periodic check at workers to see if the memory limit has been
exceeded. We implement this by having the worker control thread send
a message to each domain hosted at the worker, instructing it to update
a shared atomic integer with the current aggregate size of its nodes'
states. We then compare the sum of these per-domain integers to the
global memory limit. Currently, the check is correlated with heartbeats
to the controller (once a second) for ease of implementation, but we
could easily decouple it.

Note that the state sizes read are approximations: with many domains (>>
worker threads), the probe packet will not be delivered immediately, so
recent state size increases may not be reflected in the sizes read until
the next check occurs (~1 sec later).

In the future, we'll probably want a more advanced version of this that
records information on a per-node basis.