Evict a target number of bytes, not keys
Changes the high-level eviction logic to instruct the largest domain to
evict a target number of bytes, rather than a fixed number of keys. When
handling the eviction, the domain keeps evicting keys (one at a time,
for now) until it meets the target number of bytes freed. The evictions
are randomly spread across different indices.

The number of bytes is capped at the total size of the domain's partially
materialized state, to avoid live-locking the eviction. De facto, the
net effect is that more than the requested number of bytes get freed,
since downstream nodes and domains also evict the keys for which we
triggered an eviction. However, it's difficult to estimate the
transitive effect of an eviction, so we don't try.

evict-o-rama now manages to run within a 100 KB memory budget!