Enforce primary key uniqueness constraint
This changes the semantics of Soup's primary keys to make them unique.
Specifically, a base with a primary key set that receives a write for
a specific key now checks that it does not already have a record for
this key. If it does, it silently discards the write.

A few things to note:
 1. The client never finds out about the discarded write; Soup
    provides no acknowledgement whether all writes in a batch succeeded.
 2. This breaks the `update` method on `Mutator`, which injects a batch
    consisting of a delete followed by an insert. Since we process the
    entire batch at the base *before* applying changes to the
    materialized state, the insert gets filtered out by the primary key
    checking logic. We can fix this by having an `update` operation on
    base nodes, and a more expressive `update` interface on mutators.
 3. We don't currently support "replace" semantics (à la SQL "REPLACE
    INTO"), though it's relatively straightforward to add them.

With this change, Django/JConf login and session management now work,
and the application is usable while running on Soup!