Run setup queries one at a time
For silly reasons, Soup and the MySQL shim don't deal well with multiple
queries arriving in a single statement, so we can't run all of
db-schema.sql as one query.

The fix is inelegant, but I don't see immediately how the futures-based
`.and_then` chaining could be used here. A few things I tried didn't
work, so this will have to do.