Load generators now quit early if clients are slow
With this change, generators monitor how quickly clients are draining
queued jobs, and stop issuing jobs when they detect that clients have
enough queued work to last for the remaining duration of the experiment.
This is mostly a work-around for rayon-rs/rayon#544.

Note that the load generator now runs *in* the thread pool, so the
`threads` argument should now be set to the total number of cores rather
than #core - #generators. This is due to rayon-rs/rayon#562. It's a
little unfortunate because it means that *all* job distribution requires
stealing (the generator will put all jobs on its local queue).

Note also that (because of the same linked rayon issue) the creation of
`id_rng` is now in a closure. This is so that the argument can be `Send`
so we can get it into the thread pool in the first place.