Proper recipe expression parsing
Replaces the ad-hoc string processing with a parser that can handle a
more systematic set of specifications:

 1. Named, public views: [VIEW|QUERY] name: SELECT ...
 2. Named, internal views: name: SELECT ...
 3. Unnamed, public views: [VIEW|QUERY]: SELECT ...
 4. Unnamed, internal views: SELECT ...

(Technically, 4. allows ": SELECT ...", too.)

For unnamed queries, Soup will choose a name and return it to the caller
when the recipe is installed.

We need this facility to deduplicate ad-hoc queries issued by Lobsters
without having to run a migration every time. (Currently, we perform
>100 migrations for entirely identical queries, only to notice that
there's nothing to do.)