Make secondary indices work with non-unique primary indices
Previously secondary indices would point to just the primary index
value, without the KeyIndex or the KeySeq. This was correct for unique
primary indices, but not otherwise. Consider the following scenario:

Table (a, b)
Indices: [1], [0]

Values:
\# a  b
1  0  0
2  0  1
3  0  0
4  0  1

.lookup([1], values [0]) would here retrieve
the primary key values for rows 1 and 3, and then try to follow those
pointers with a call to .lookup([0], values [0]). This would return all
4 rows - twice.

What we're really interested in here is to simply find
the values for rows 1 and 3, which we do as long as we include the full
key in the secondary index. Instead of the second look up we'll then do
something like .get(serialized version of (0, [0], seq number)).