tokio reactors are *no longer* Send
And so we can't actually do the unsafe impl Send for mssql::Conf.
Specifically, the handle given to the sql connection is constructed on a
different thread than where it ends up, and that causes the connection
to simply stop progressing, even though we drive that reactor. This is
probably related to the whole tokio reform thread pool business.