Swap before checking if key is still missing
This is important, as a reader may request a key just *after* we
backfilled that key and swapped, but *before* we swapped again. Since
writes are applied to state only at the beginning of a swap, the state
we get back from the first swap does not contain the key that we just
filled, and so we'll do the reply a second time!