Add global predicates to data-flow
This adds support for generating filters for global predicates to the
query planner. This allows us to correctly place both disjunctive (OR)
conditionals and filters on computed columns (e.g., aggregations and the
results of arithmetic between columns, and between columns and
literals).

Supporting global predicates requires some care: for example, we must
ensure that we compute any columns that the predicates depend on
upstream of the filter node inserted.

With this change, we can now correctly lay out the data-flow nodes for
the complex notification query in Lobsters, although bugs remain when
trying to actually run the data-flow.

In the future, I want to rewrite the query planner; it's become too deep
a stack of hacks made for deadlines.