Check that we don't run into newfound bug
Specifically, consider the case where you have an aggregation that
groups by [0,1], and a downstream reader keyed by [0]. The reader
will be partial, and add an index on [0] on the aggregation. When the
reader misses, it will ask the aggregation for a replay of, say, ["7"].
The aggregation will perform that replay, but when trying to process the
response, it will miss when doing a self-lookup of ["7", x] for whatever
value x happens to be in column 1 of the replay responses. This triggers
a second round of (unnecessary) replays, which in and of itself causes a
bunch of issues, but let's imagine that it works fine, and that both
["7"] and ["7", x] is filled correctly in the aggregation. Now a write
comes along for ["7", y]. The aggregation misses on its lookup, so the
write is discarded. But the reader holds state for ["7"], so this
violates key monotonicity!

There are a couple of ways that this could be fixed. One idea is that
the aggregation could evict ["7"] on miss, similar to what we currently
do for joins. This would cause us to do a lot more replays, and will
likely reveal a bunch of bugs with overlapping replays, but should be
theoretically correct. An alternative, and more robust, idea is to to
have Soup realize that [0,1] and [0] overlap, and that the replay of [0]
actually fills *all* holes in [0,1].

Separately, the *reason* why we run into this situation in the majority
of cases is that Soup has to pull through columns for downstream views
(e.g., consider the case where 0 = article id and 1 = title). In these
cases, there is *really* just one group by key (article id), and the
other column value is uniquely determined by the first. In this case,
the aggregation operator (and other similar operators) could be modified
to know about these kinds of "pull through" columns, and simply pick any
1 value for a set of records sharing the same value for [0]. This is
similar to what MySQL does when it is lax about not having to name all
selected columns in a GROUP BY (PostgreSQL does not allow this).