Project lifted arith./lit. columns into output
If the query asks to project an arithemetic or literal column that we
compute early in order to use it in a global predicate, it needs to be
projected into the output as an "ordinary" column.

Fixes the strange "tried to emit column not in getter" crash in the
Lobsters benchmark.