Top-K is buggy
We've found a bug in Top-K that means it can keep and return more than k
elements in some cases (in particular, when elements not previously in
the Top-K enter the Top-K collection, and when these changes are
triggered by positive/negative pairs from bases).

Adds a test case that exercises this behavior.