Configurable eviction frequency
Introduces a configuration option to control the frequency at which
we check the memory limit and trigger evictions if we exceed it.

For Lobsters at 1,000 page views/sec, we need a check frequency of 10ms
to keep up with the state growth. However, since all interaction
involved is local to a worker, this is fine. If we wanted hard memory
limits, or need to deal with even faster rates of growth, we'll probably
have to come up with a new way of doing things.