Fix join reordering bug
Since we picked join table ordering based on the order in which tables
appeared in the join condition (!), we would sometimes reverse the order
of ancestor nodes in a join. This doesn't matter for normal joins, but
is very much incorrect for left joins.

Apparently past me had the foresight to smell something fishy and left
an XXX comment, but unfortunately didn't actually realize the problem at
the time.