Remove local optimization for external api
We basically never expect *clients* to be in the same process as the
server, so this optimization is unused, but adds significant complexity.
We *will* want to use it internally between domains on the same machine,
but that's all internal to the dataflow, and not something the api needs
to know about.