Avoid duplicating node assignments in DomainHandle
This avoids keeping state about node assignment to domains in two places
(the graph and DomainHandle), at the cost of doing a graph traversal to
identify the nodes in a domain.

N.B.: makes recovery O(|D| * |V|), where D is the set of domains, and V
is the set of vertices!