Incomplete attempt to fix recursive node removal
The reason why we end up with leftover migration state is that the
previous node removal code overzealously removed parent nodes, even if
they belong to an internal view (e.g., "VoteCount"). Such an internal
view can be unaffected by a failure, and should survive. Either way,
blowing away its nodes without clearing other state is a recipe for
disaster.

This fix tries to amend the code such that it only removes parent nodes
until it hits either (i) a base, (ii) a query leaf that corresponds to
an internal view different to the one associated with the removed leaf
node, or (iii) a node that still has other children. This part works,
but we now in some settings fail to remove the internal view even if it
*did* fail: I believe we must process removals in topological graph
order. Ugh.

Note: the utility method `SqlIncorporator::is_leaf_address` added by
this commit is O(N) in the number of queries in the graph; ideally, it'd
be O(1).