Revert "Attempt test harness for ControllerInner"
This reverts commit 0c405efb0e208a23a1c47abd43107441ae425d49
and 54b6158304f313c1dcf196644916b336f4bbda03, which tried to add a test
harness that turned out to be too complex (for now, at least). I'm
leaving the commits in the history for merging in case we want to
resurrect the `MockWorker` test framework for `ControllerInner` in the
future.