Note ingress reuse bug for future reference
It's not okay to re-use any egress/ingress between a pair of domains,
since the egress/ingress nodes may be sharded by a different column than
those we would otherwise add.

Found this in passing, and remembered that we discussed it before. Not
the cause of my bug, but worth flagging so that we resolve it.