Fix nasty partial backfill bug with sharding
This fixes the bug observed by RTM's example program. The problem was
that we didn't correctly handle domains receiving replays they never
requested. This can sometimes happen for legitimate reasons in
moderately complex graphs: for example, a partial vertex may query back
into unsharded graph regions with a replay key different to its sharding
key. When processing the backfills forwards, the sharder will send them
to *all* shards, since it has no state to indicate which shard requested
the backfill to begin with (replay paths are per-node, not per-shard).
Consequently, some shards receive backfills they do not recall requesting
(and thus aren't waiting for).

Previously, we would crash on receiving such backfills, although due to
buggy checking logic, we would sometimes continue processing them if the
shard *happened* to be waiting for a backfills from the same source (but
for a different key). Instead, we now (correctly) ignore these backfills.
Since backfills are batched, this requires pruning each incoming partial
`ReplayPiece` to remove keys that we're not waiting for (and their
records).