Fix up test for RTM's sharding bug
This test previously triggered the sharding bug with a simplified
graph. It's a slight modification of the original graph: due to another
issue, which we'll fix separately.

The issue is this: emitting a join column twice (as real queries
sometimes do, and we do for convenient reasoning at higher, relational
layers) and resharding on the "copy" confuses our logic. Soup correctly
infers that there is no shuffle and only needs to ask one shard (as the
columns are equivalent), but the graph planning code, which isn't as
smart, inserts a resharding, which waits for replay responses from *all*
shards.