It's NOT okay to alias output colums of views
Fixes a bug introduced by 921840955: we were declaring the column of
each view to be an alias of (i.e,, to equivalent to) the input
column.

The input column carries as its table name the base table or view from
which it derives (i.e., the table or operator that "introduces" the
column), but it often isn't actually *equivalent* to said
table/operator's column of the same name. Consider a column passed
through the GROUP BY of an aggregation: clearly, it is no longer the
same as the original column, since duplicate values have been removed.

Since the definition of `mir::Column::aliases` is that the column is in
all respects *equivalent* to its alias columns, we may only add aliases
when all rows are preserved and the columns are indeed row-by-row
equivalent (e.g., join columns).