Workaround for TopK column equality in Lobsters
In certain situations, TopK nodes at the leaves of our queries can end
up with columns whose `function` fields have been cleared. Since we
compared `function` when comparing `mir::Column` objects for equality,
this led to problems when the parent vertex outputs a function column
(as with aggregations).

With this change, we compare columns *only* by name and table. I believe
that this is always correct, even though it's slightly less strict. We
might be able to fix the TopK case so that we can revert this, but for
now this should be sufficient to get Lobsters to run again.