Trace view columns to determine types for schema
Work towards returning the right column type with `View::schema`.
Previously, we hard-coded the type to `SqlType::Text` (and before that,
the shim had to maintain the information). This implementation traces
the view column back to a base table (or another intermediate view) and
obtains the type from the result column. Works for base types; the
implemention for computed columns and projected literals is a bit
finicky and not done yet.