Panic if existing query redefined
Be more careful about how we handle repeated user-specified query names.
They're fine as long as the query is identical, but not if it is not --
the name would map to the new query and leave the existing one orphaned.
We therefore disallow redefining queries without a migration that first
removes the old query, at least for the moment.