Don't send writes to nodes beyond mat. frontier
This patch makes two main changes: First, egress nodes will no longer
send (or even consider) sending writes to child nodes that are marked as
being beyond the materialization frontier. Second, security unions are
now tracked as such, and are always placed in their own domain.
Together, these make it so that if everything past a security union is
marked as being beyond the materialization frontier, egress nodes will
not send writes to them. This has the nice benefit that writes are no
longer sent to all security universes if they don't have to be (and
should thus no longer need to be linear in the number of universes)!