HACK: Force shallow for piazza sec agg
This makes it so that the aggregation that combines the post counts
after the security union in piazza with basic policies enabled is marked
as shallow. Normally we could do this with `CREATE VIEW SHALLOW_`, but
for some reason the code chokes on that, so we do this for now. It's
ugly, and doesn't work for complex policies/any other policies.