Avoid blocking the reactor when processing
When we process a packet, we want to do so under the `blocking` label so
that we don't hold up the tokio reactor while processing. However, if
there are many blocking threads (100 by default), the call to `blocking`
can fail. In the old version of the code (and still in some places), we
simply `.wait()` until we're free to block. This is not generally a good
idea when you have many threads, because it may mean that we hit the
limit and _all_ the remaining threads are stopped spinning to become
blocking threads. This change fixes the most egregious part of the code
that did this: packet processing. It will now back off of the current
packet if it can't block, and will thus be freed up to handle other
requests in the meantime.