Refactored all my filteraggregate changes to use the new syntax Malte created in nom-sql.
Also made yet another attempt at reindexing based on a conversation
with Malte that made me realize the filter indices are based on the
parent's columns. It doesn't seem to be working yet, but it really
should be a separate commit anyway, I just started working on it in
this one because I forgot I had this half-finished uncommitted code
from yesterday.

Anyway, it compiles, hallelujah, and it's past my bedtime so debugging
sounds like a problem for the morrow. Also, don't you think these
commit messages would look cooler with the Star Wars opening scroll?