Make remove_leaf search better for reader
The function is given the `NodeIndex` of a node that _has_ a `Reader`
attached. That Reader may be deeper down, such as past a shuffle or a
domain boundary, which the old code did not catch. Furthermore, in the
case where we're removing the reader leaf, it's fine if the node that
was initially given to `remove_leaf` itself has other children. The code
below then just won't remove that.